app/requirements.txt
Flask==3.1.*
Flask-HTTPAuth==4.8.*
jsonpickle==4.1.*
paho-mqtt== 2.1.*
pydantic==2.11.*
python-dotenv==1.1.*
requests==2.32.*
PyYAML==6.0.*
xxtea==3.3.*
astral==3.*
tzlocal==5.3.*
